feat: consolidate avatar rigging and enable instanced skinning
